Remove bang on update and reorganise interactors into two factor namespace
